app/requirements.txt
streamlit
sqlalchemy
psycopg2-binary
pandas
plotly
openai
python-dotenv
